refactor: `calculate` operand 리팩터링
