+++ src/constants/alertMessage.jsx
@@ -1,0 +1,7 @@
+const ALERT = {
+  INFINITY: '오류입니다. AC를 눌러 값을 초기화해주세요.',
+  MAX_DIGIT: '숫자는 한번에 최대 3자리 수까지 입력 가능합니다.',
+  OPERATOR_EXIST: '=을 눌러 계산해주세요',
+};
+
+export default ALERT;
