+++ src/constants/buttonChar.jsx
@@ -1,0 +1,2 @@
+export const digits = ['9', '8', '7', '6', '5', '4', '3', '2', '1', '0'];
+export const operations = ['/', 'X', '-', '+', '='];
