+++ src/index.jsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import ReactDOM from 'react-dom/client';
 import './css/index.css';
 import reportWebVitals from './js/reportWebVitals';
-import Calculator from './Calculator';
+import Calculator from './components/Calculator';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
